requirements.txt

ninja==1.11.1.1
numpy==1.24.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
openpyxl==3.1.5
pandas==2.0.3
pillow==10.4.0
pyparsing==3.1.2
pytz==2024.1
PyYAML==6.0.2
rarfile==4.2
regex==2024.7.24
requests==2.32.3
robustbench==1.1
safetensors==0.4.4
scikit-learn==1.3.2
scipy==1.10.1
screen==1.0.1
seaborn==0.13.2
sympy==1.13.2
tencentcloud-sdk-python==3.0.1225
threadpoolctl==3.5.0
timm==0.3.2
torch==2.4.0
torchaudio==2.4.0
torchsde==0.2.6
torchvision==0.19.0
tqdm==4.66.5
trampoline==0.1.2
triton==3.0.0
ttach==0.0.3

